+++ R/calc_sun_rise_set.R
@@ -1,23 +1,25 @@
 #' Calculates the time of sunrise and sunset
 #' 
 #' Calculates the time of sunrise and sunset based on latitude and date.
-#'
-#' @param datetimes Vector of dates as \code{POSIXct} or \code{POSIXlt} (see \code{\link{DateTimeClasses}}) format, but in SOLAR time
-#' @param lat Single latitude value of site. South should be negative, north positive
-#' @return data.frame of 
+#' 
+#' @param date Vector of dates in \code{Date} format.
+#' @param latitude Single latitude value of site. South should be negative, 
+#'   North positive
+#' @return data.frame of sunrise and sunset (apparent solar time, nominally UTC)
 #' @importFrom LakeMetabolizer sun.rise.set
+#' @importFrom lubridate force_tz
 #' @keywords internal
 #' @examples
-#' calc_sun_rise_set(lat=40.75,datetimes=as.POSIXlt('2013-03-31'))
-#' @author
-#' Luke A. Winslow
-#' @seealso 
-#' \link{calc_is_daytime}
+#' calc_sun_rise_set(latitude=40.75, 
+#'   date=as.POSIXlt(c('2013-03-31', '2017-07-01')))
+#' @seealso \link{calc_is_daytime}
 #' @export
-calc_sun_rise_set <- function(datetimes, lat) {
-  sun.rise.set <- LakeMetabolizer::sun.rise.set(datetimes, lat)
+calc_sun_rise_set <- function(date, latitude) {
+  app.solar.time <- as.POSIXct(strftime(date, "%Y-%m-%d 0"), format="%Y-%m-%d %H") # use local tz, as LakeMetabolizer does
+  sun.rise.set <- LakeMetabolizer::sun.rise.set(app.solar.time, latitude)
   # reformat
-  data.frame(sunrise=sun.rise.set[,1], sunset=sun.rise.set[,2])
+  data.frame(sunrise=lubridate::force_tz(sun.rise.set[,1], tz="UTC"), 
+             sunset=lubridate::force_tz(sun.rise.set[,2], tz="UTC"))
 }