+++ R/mm_is_valid_day.R
@@ -30,7 +30,7 @@ mm_is_valid_day <- function(day, day_start=4, day_end=27.99,
   stop_strs <- character(0)
   
   # estimate time steps - useful for a few tests
-  timestep.days <- if(is.na(timestep_days)) suppressWarnings(mean(as.numeric(diff(v(day$local.time)), units="days"), na.rm=TRUE)) else timestep_days
+  timestep.days <- if(is.na(timestep_days)) suppressWarnings(mean(as.numeric(diff(v(day$solar.time)), units="days"), na.rm=TRUE)) else timestep_days
   timestep <- as.difftime(timestep.days, units="days")
   if(!is.finite(timestep)) {
     stop_strs <- c(stop_strs, "can't measure timesteps")
@@ -39,21 +39,21 @@ mm_is_valid_day <- function(day, day_start=4, day_end=27.99,
   # Require that the data span the full expected period (e.g., from 10:30pm on
   # preceding day to 6am on following day)
   if('full_day' %in% tests & is.finite(timestep)) {
-    date_counts <- table(format(day$local.time, "%Y-%m-%d"))
-    local_date <- names(date_counts)[which.max(date_counts)]
-    date_start <- as.POSIXct(paste0(local_date, " 00:00:00"), tz=lubridate::tz(v(day$local.time)))
+    date_counts <- table(format(day$solar.time, "%Y-%m-%d"))
+    ply_date <- names(date_counts)[which.max(date_counts)]
+    date_start <- as.POSIXct(paste0(ply_date, " 00:00:00"), tz=lubridate::tz(v(day$solar.time)))
     similar_time <- function(a, b, tol=timestep) {
       abs(as.numeric(a, units="days") - as.numeric(b, units="days")) < as.numeric(tol, units="days")
     }
-    if(!similar_time(min(day$local.time)-date_start, day_start))
+    if(!similar_time(min(day$solar.time)-date_start, day_start))
       stop_strs <- c(stop_strs, "data don't start when expected")
-    if(!similar_time(max(day$local.time)-date_start, day_end))
+    if(!similar_time(max(day$solar.time)-date_start, day_end))
       stop_strs <- c(stop_strs, "data don't end when expected")
   }
   
-  # Require that on each day local.time has a ~single, ~consistent time step
+  # Require that on each day solar.time has a ~single, ~consistent time step
   if('even_timesteps' %in% tests & is.finite(timestep)) {
-    timestep.deviations <- suppressWarnings(diff(range(as.numeric(diff(v(day$local.time)), units="days"), na.rm=TRUE)))
+    timestep.deviations <- suppressWarnings(diff(range(as.numeric(diff(v(day$solar.time)), units="days"), na.rm=TRUE)))
     if(!is.finite(timestep.deviations)) {
       stop_strs <- c(stop_strs, "can't measure range of timestep lengths")
     } else {