+++ tests/testthat/test-metab_helpers.R
@@ -5,18 +5,18 @@ test_that("mm_data works", {
   # runs and can be used to select columns
   expect_is(mm_data(), "unitted_data.frame")
   expect_equal(nrow(mm_data()), 1)
-  expect_equivalent(mm_data(local.time), mm_data()["local.time"])
-  expect_equivalent(mm_data(depth, temp.water, local.time), mm_data()[c("depth","temp.water","local.time")])
+  expect_equivalent(mm_data(solar.time), mm_data()["solar.time"])
+  expect_equivalent(mm_data(depth, temp.water, solar.time), mm_data()[c("depth","temp.water","solar.time")])
   
   # 'optional' attribute is set sensibly
   expect_equal(attr(mm_data(NULL), 'optional'), 'all')
-  expect_equal(attr(mm_data(depth, temp.water, local.time), 'optional'), 'none')
-  expect_equal(attr(mm_data(local.time, DO.obs, optional='DO.obs'), 'optional'), 'DO.obs')
-  expect_equal(attr(mm_data(local.time, DO.obs, optional=c('DO.obs','local.time')), 'optional'), 'all')
+  expect_equal(attr(mm_data(depth, temp.water, solar.time), 'optional'), 'none')
+  expect_equal(attr(mm_data(solar.time, DO.obs, optional='DO.obs'), 'optional'), 'DO.obs')
+  expect_equal(attr(mm_data(solar.time, DO.obs, optional=c('DO.obs','solar.time')), 'optional'), 'all')
 
   # 'optional' attribute is checked
-  expect_error(mm_data(local.time, DO.obs, optional='DO.sat'), "'arg' should be one of")
-  expect_error(mm_data(local.time, DO.obs, optional=c('DO.obs','all')), "should be length 1")
+  expect_error(mm_data(solar.time, DO.obs, optional='DO.sat'), "'arg' should be one of")
+  expect_error(mm_data(solar.time, DO.obs, optional=c('DO.obs','all')), "should be length 1")
 })
 
 test_that("mm_validate_data works", {
@@ -43,7 +43,7 @@ test_that("mm_validate_data works", {
   library(unitted)
   ok_data <- eval(formals(metab_mle)$data)
   expect_error(mm_validate_data(NULL, NULL, "metab_mle"), "data is NULL but required")
-  expect_error(mm_validate_data(data.frame(), NULL, "metab_mle"), "data is missing these columns: local.time, DO.obs, DO.sat, depth, temp.water, light")
+  expect_error(mm_validate_data(data.frame(), NULL, "metab_mle"), "data is missing these columns: solar.time, DO.obs, DO.sat, depth, temp.water, light")
   expect_error(mm_validate_data(dplyr::select(ok_data,-temp.water), NULL, "metab_mle"), "data is missing these columns: temp.water")
   expect_error(mm_validate_data(dplyr::mutate(ok_data,temp.air=9), mm_data('temp.air'), "metab_mle"), "data should omit these extra columns: temp.air")
   expect_error(mm_validate_data(dplyr::mutate(ok_data,temp.water=u(temp.water, "degF"), DO.obs=u(DO.obs, "mgO L^-1")), NULL, "metab_mle"), "unexpected units in data:")
@@ -65,8 +65,8 @@ test_that("mm_is_valid_day works", {
   library(unitted)
   french <- streamMetabolizer:::load_french_creek()
  
-  good_day <- u(dplyr::filter(v(french), local.time >= as.POSIXct("2012-08-24 22:30:00", tz="Etc/GMT+7"),
-                              local.time <= as.POSIXct("2012-08-26 06:00:00", tz="Etc/GMT+7")), get_units(french))
+  good_day <- u(dplyr::filter(v(french), solar.time >= as.POSIXct("2012-08-24 22:30:00", tz="UTC"),
+                              solar.time <= as.POSIXct("2012-08-26 06:00:00", tz="UTC")), get_units(french))
   bad_day <- dplyr::mutate(
     good_day,
     DO.obs=replace(DO.obs, 40, u(NA, get_units(DO.obs))),
@@ -78,8 +78,8 @@ test_that("mm_is_valid_day works", {
   expect_true(mm_is_valid_day(good_day, day_start=-1.5, day_end=30))
   
   # test faulty timestep
-  dateless_day <- good_day; dateless_day$local.time <- replace(dateless_day$local.time, 2:(nrow(dateless_day)-1), NA)
-  expect_equal(mm_is_valid_day(dateless_day), c("can't measure timesteps", "NAs in local.time"))
+  dateless_day <- good_day; dateless_day$solar.time <- replace(dateless_day$solar.time, 2:(nrow(dateless_day)-1), NA)
+  expect_equal(mm_is_valid_day(dateless_day), c("can't measure timesteps", "NAs in solar.time"))
   
   # test full_day
   expect_equal(mm_is_valid_day(good_day, day_start=-10, day_end=30), "data don't start when expected")
@@ -100,18 +100,18 @@ test_that("mm_filter_valid_days works", {
   library(dplyr); library(unitted)
   french <- streamMetabolizer:::load_french_creek()
   
-  french_daily <- data.frame(local.date=as.Date(sprintf("2012-08-%2d",15:30)), K600=7)
+  french_daily <- data.frame(date=as.Date(sprintf("2012-08-%2d",15:30)), K600=7)
   
   expect_equal(length(mm_filter_valid_days(french, day_start=10, day_end=12)), 3)
-  expect_equal(nrow(mm_filter_valid_days(french, data_daily=french_daily, day_start=10, day_end=12)[[1]]), 700)
+  expect_equal(nrow(mm_filter_valid_days(french, data_daily=french_daily, day_start=10, day_end=12)$data), 672)
 })
 
 test_that("mm_filter_dates works", {
   
   start_time <- Sys.time()
   start_date <- as.Date(start_time)
-  udat <- data.frame(local.time=start_time + as.difftime(1:100, units='hours'), value=1:100)
-  ddat <- data.frame(local.date=start_date + as.difftime(1:100, units='days'), value=1:100)
+  udat <- data.frame(solar.time=start_time + as.difftime(1:100, units='hours'), value=1:100)
+  ddat <- data.frame(date=start_date + as.difftime(1:100, units='days'), value=1:100)
   # no filter with defaults
   expect_equal(streamMetabolizer:::mm_filter_dates(udat), udat)
   expect_equal(streamMetabolizer:::mm_filter_dates(ddat), ddat)