+++ NEWS.md
@@ -1,3 +1,13 @@
+# streamMetabolizer 0.8.0
+
+## Changes
+
+* Major interface change (renamed variable) to clarify types of time: solar.time
+(mean solar time), app.solar.time (apparent solar time), local.time (time in 
+local time zone). Metabolism models now accept solar.time rather than 
+local.time, though it's still possible to pass in local time but just call it
+solar.time (as long as you don't have daylight savings time).
+
 # streamMetabolizer 0.7.3
 
 ## Changes